Add files via upload
This is a sample of a store management system which allows you to store the items that you sell in a database. It  comes with a Point Of Sale (POS) which allows you to do transactions and record them in the database . You can also make changes to data in the database via the update page. I have also added code that enables you to print a receipt after a transaction has been completed. It  needs a bit of improvement especially on the front end. Feel free to contribute and thanks in advance.